+++ src/functions/public/Expand-AccountPermission.ps1
@@ -21,7 +21,7 @@ function Expand-AccountPermission {
         # Object that was output from Format-SecurityPrincipal
         $AccountPermission,
 
-        # Properties to exclude from the output
+        # Properties to exclude from the output because they cause problems or are unnecessary/redundant/undesirable
         # All properties listed on a single line to workaround a bug in PlatyPS when building MAML help
         # (error is 'Invalid yaml: expected simple key-value pairs')
         # Caused by multi-line default parameter values in the markdown