+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.0.62] - 2024-01-13 - bug fix in New-NtfsAclIssueReport
+
+## [2.0.61] - 2024-01-13 - changed new-ntfsaclissuereport to not depend no the NtfsAccessControlEntries property of the $UserPermissions object
+
+## [2.0.60] - 2024-01-13 - changed new-ntfsaclissuereport to not depend no the NtfsAccessControlEntries property of the $UserPermissions object
+
 ## [2.0.59] - 2024-01-13 - bug fix in expand-accountpermission
 
 ## [2.0.58] - 2024-01-12 - bugfix in convertto-simpleproperty on lines 48-51